requirements.txt
altgraph==0.17.3
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0ek9yztvu3/croot/certifi_1665076692562/work/certifi
click==8.1.3
docopt==0.6.2
macholib==1.16.2
num2words==0.5.12
Pillow==9.5.0
pyinstaller==5.6.1
pyinstaller-hooks-contrib==2022.11
PyQt5==5.15.9
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.4.3.2
python-dotenv==1.0.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
textwrap3==0.9.2
